requirements.txt
aiohttp>=3.9
anthropic>=0.40
python-dotenv>=1.0
